// projects/radui-library/src/lib/table/table-interface.ts
export interface TableInterface {
    name: string;
    ColumnDefinition: [];
    columnID: any[];
    header: any[];
    dataObj: any;
    rowData: any[];
    headerColor: string;
    headerFontColor: string;
    backgroundColor: string;
    headerFontSize: string;
    borderColor: string;
    borderStyle: string;
}